# Conflicts:
#	random_number.uml
#	src/main/java/de/cae/Generator/PolarGenerator.java
#	src/main/java/de/cae/Guete/Autokorrelation.java
#	src/main/java/de/cae/Guete/SequenzTest.java

+++ src/main/java/de/cae/Generator/GeneratorFactory.java
@@ -3,18 +3,21 @@ package de.cae.Generator;
 /**
  * This Factory provides Generators that are pre Generated or stubs to generate your own one.
  * The provided pre generated Generators are:
- * - ANSI-C
- * - Minimal Standard
- * - RANDU
- * - SIMSCRIPT
- * - NAG's LCG
- * - Maple's LCG
+ * <ul>
+ *     <li>ANSI-C</li>
+ *     <li>Minimal Standard</li>
+ *     <li>RANDU</li>
+ *     <li>SIMSCRIPT</li>
+ *     <li>NAG's LCG</li>
+ *     <li>Maple's LCG</li>
+ * </ul>
  * <p>
  * The used Generators are:
- * - {@link LCG}
- * - {@link PolarGenerator}
- * - {@link SinusGenerator}
- *
+ * <ul>
+ * <li>{@link LCG}</li>
+ * <li>{@link PolarGenerator}</li>
+ * <li>{@link SinusGenerator}</li>
+ * </ul>
  * @author Cevin Voigt
  * @version 1.0
  */
@@ -54,80 +57,81 @@ public class GeneratorFactory {
 
     /**
      * Generates an {@link LCG Linaer Congruential de.cae.Generator} with the Values:
-     * - Multiplier = 1103515245
-     * - Shift = 12345
-     * - Modulator = 2^31
-     * - First Value = 12345
+     * <br>
+     * - Multiplier = 1103515245<br>
+     * - Shift = 12345<br>
+     * - Modulator = 2^31<br>
+     * - First Value = 12345<br>
      *
      * @return An {@link LCG}
      */
     public static LCG AnsiC() {
-        return getLCG(1103515245, 12345, (int) Math.pow(2, 31), 12345);
+        return getLCG(1103515245, 12345, (long) Math.pow(2, 31), 12345);
     }
 
     /**
-     * Generates an {@link LCG Linaer Congruential de.cae.Generator} with the Values:
-     * - Multiplier = 16807
-     * - Shift = 0
-     * - Modulator = 2^31 - 1
-     * - First Value = 1
+     * Generates an {@link LCG Linaer Congruential de.cae.Generator} with the Values:<br>
+     * - Multiplier = 16807 <br>
+     * - Shift = 0<br>
+     * - Modulator = 2^31 - 1<br>
+     * - First Value = 1<br>
      *
      * @return An {@link LCG}
      */
     public static LCG MinimalStandard() {
-        return getLCG(16807, 0, ((int) Math.pow(2, 31)) - 1, 1);
+        return getLCG(16807, 0, ((long) Math.pow(2, 31)) - 1, 1);
     }
 
     /**
-     * Generates an {@link LCG Linaer Congruential de.cae.Generator} with the Values:
-     * - Multiplier = 65539
-     * - Shift = 0
-     * - Modulator = 2^31
-     * - First Value = 1
+     * Generates an {@link LCG Linaer Congruential de.cae.Generator} with the Values:<br>
+     * - Multiplier = 65539<br>
+     * - Shift = 0<br>
+     * - Modulator = 2^31<br>
+     * - First Value = 1<br>
      *
      * @return An {@link LCG}
      */
     public static LCG RANDU() {
-        return getLCG(65539, 0, ((int) Math.pow(2, 31)), 1);
+        return getLCG(65539, 0, ((long) Math.pow(2, 31)), 1);
     }
 
     /**
-     * Generates an {@link LCG Linaer Congruential de.cae.Generator} with the Values:
-     * - Multiplier = 630360016
-     * - Shift = 0
-     * - Modulator = 2^31 - 1
-     * - First Value = 1
+     * Generates an {@link LCG Linaer Congruential de.cae.Generator} with the Values:<br>
+     * - Multiplier = 630360016<br>
+     * - Shift = 0<br>
+     * - Modulator = 2^31 - 1<br>
+     * - First Value = 1<br>
      *
      * @return An {@link LCG}
      */
     public static LCG SIMSCRIPT() {
-        return getLCG(630360016, 0, ((int) Math.pow(2, 31)) - 1, 1);
+        return getLCG(630360016, 0, ((long) Math.pow(2, 31)) - 1, 1);
     }
 
     /**
-     * Generates an {@link LCG Linaer Congruential de.cae.Generator} with the Values:
-     * - Multiplier = 13^13
-     * - Shift = 0
-     * - Modulator = 2^59
-     * - First Value = 123456789
+     * Generates an {@link LCG Linaer Congruential de.cae.Generator} with the Values: <br>
+     * - Multiplier = 13^13 <br>
+     * - Shift = 0 <br>
+     * - Modulator = 2^59 <br>
+     * - First Value = 123456789 <br>
      *
      * @return An {@link LCG}
      */
     public static LCG NAGsLCG() {
-        return getLCG((int) Math.pow(13, 13), 0, ((int) Math.pow(2, 59)), 123456789);
+        return getLCG((long) Math.pow(13L, 13L), 0, ((long) Math.pow(2, 59)), 123456789);
     }
 
     /**
-     * Generates an {@link LCG Linaer Congruential de.cae.Generator} with the Values:
-     * - Multiplier = 427419669081L
-     * - Shift = 0
-     * - Modulator = 10^12 - 11
-     * - First Value = 1
+     * Generates an {@link LCG Linaer Congruential de.cae.Generator} with the Values: <br>
+     * - Multiplier = 427419669081 <br>
+     * - Shift = 0 <br>
+     * - Modulator = 10^12 - 11 <br>
+     * - First Value = 1 <br>
      *
      * @return An {@link LCG}
      */
     public static LCG MaplesLCG() {
-        return getLCG(427419669081L, 0, ((int) Math.pow(10, 12)) - 11, 1);
+        return getLCG(427419669081L, 0, ((long) Math.pow(10, 12)) - 11, 1);
     }
 
 }
